Update to new release
This also removes cffi from a runtime dependency (but still a build
dependency). This is done because there's a c-extension that is used by
default, most users of conda aren't using pypy where the cffi module
would make sense, and those users that want the cffi version can install
cffi themselves as an optional dependency to use this.